+++ util.c
@@ -1891,7 +1891,7 @@ static bool stratum_benchdata(json_t *result, json_t *params, int thr_id)
 	json_object_set_new(val, "freq", json_integer((uint64_t)cpufreq));
 	json_object_set_new(val, "memf", json_integer(0));
 	json_object_set_new(val, "power", json_integer(0));
-	json_object_set_new(val, "khashes", json_real((double)global_hashrate / 1000.0));
+	json_object_set_new(val, "khashes", json_real(global_hashrate / 1000.0));
 	json_object_set_new(val, "intensity", json_real(opt_default_priority));
 	json_object_set_new(val, "throughput", json_integer(opt_n_total_threads));
 	json_object_set_new(val, "client", json_string(PACKAGE_NAME "/" PACKAGE_VERSION));
